Refactoring hiding classes writing comments ------------------- Schematic editor file also stores all files in symbol editor copy pasted that template with just libname change now KiCad is able to display pins correctly in schematic editor. --------------------- alignment point found
